End-to-end publish test using examples from docs (#199)
